# openapi.yaml
...
openapi: 3.2.0
info:
  title: Recension API
  summary: Continuous behavioral regression testing
  description: |-
    HTTP API for Recension: SDK clients submit versioned testcase captures; the dashboard authenticates with session tokens to browse teams, suites, batches, and promote baselines.

    **Auth**
    - Dashboard (session): `Authorization: Bearer <token>` (default TTL 7 days / `RECENSION_SESSION_TTL`)
    - SDK / client: `X-Recension-API-Key: <key>` (per-user; inherits team memberships)

    **Tenancy / RBAC**
    - Users access teams only via membership roles: `owner` | `admin` | `member` | `viewer`
    - viewer: read; member: + submit/seal; admin: + create suites, promote, manage members; owner: + grant owner
    - Cross-team access returns 404 (does not leak existence)
    - Creating a team makes the creator `owner`; bootstrap user owns seeded team `acme`

    **Hierarchy:** team → suite → batch (version/revision) → element (testcase)
  version: 0.1.0
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  contact:
    name: ayitas/recension
    url: https://github.com/ayitas/recension
servers:
- url: http://localhost:8080
  description: Local API (direct)
- url: http://localhost:3000
  description: Local gateway (nginx proxies /v1 and /healthz)
...
paths:
...
  "/v1/auth/signup":
    post:
      tags:
      - Auth
      operationId: signup
      summary: Create account
      description: Disabled when `RECENSION_ALLOW_SIGNUP=false`. Rate-limited.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AuthRequest"
      responses:
        '200':
          description: Session issued
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/SessionResponse"
        '400':
          "$ref": "#/components/responses/Error"
        '403':
          "$ref": "#/components/responses/Error"
        '409':
          "$ref": "#/components/responses/Error"
        '429':
          "$ref": "#/components/responses/Error"
  "/v1/auth/login":
    post:
      tags:
      - Auth
      operationId: login
      summary: Log in
      description: Rate-limited. Returns a signed session token (HMAC, not standard
        JWT library; same Bearer usage).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AuthRequest"
      responses:
        '200':
          description: Session issued
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/SessionResponse"
        '400':
          "$ref": "#/components/responses/Error"
        '401':
          "$ref": "#/components/responses/Error"
        '429':
          "$ref": "#/components/responses/Error"
  "/v1/auth/me":
    get:
      tags:
      - Auth
      operationId: getMe
      summary: Current user
      security:
      - SessionBearer: []
      responses:
        '200':
          description: Authenticated user (includes API key)
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/User"
        '401':
          "$ref": "#/components/responses/Error"
  "/v1/auth/api-key/rotate":
    post:
      tags:
      - Auth
      operationId: rotateApiKey
      summary: Rotate API key
      security:
      - SessionBearer: []
      responses:
        '200':
          description: User with new API key
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/User"
        '400':
          "$ref": "#/components/responses/Error"
        '401':
          "$ref": "#/components/responses/Error"
  "/v1/client/verify":
    post:
      tags:
      - Client
      operationId: clientVerify
      summary: Verify API key
      security:
      - ApiKey: []
      responses:
        '204':
          description: API key is valid
        '401':
          "$ref": "#/components/responses/Error"
  "/v1/client/submit":
    post:
      tags:
      - Client
      operationId: clientSubmit
      summary: Submit testcase captures
      description: Team/suite must already exist. Sealed batches reject further submits
        (409). Body size limited by `RECENSION_MAX_BODY_BYTES` (default 8 MiB).
      security:
      - ApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Envelope"
      responses:
        '200':
          description: Per-testcase outcomes vs baseline
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/SubmitResponse"
        '400':
          "$ref": "#/components/responses/Error"
        '401':
          "$ref": "#/components/responses/Error"
        '404':
          "$ref": "#/components/responses/Error"
        '409':
          "$ref": "#/components/responses/Error"
        '413':
          "$ref": "#/components/responses/Error"
  "/v1/blobs/{digest}":
    parameters:
    - name: digest
      in: path
      required: true
      schema:
        type: string
        pattern: "^sha256:[0-9a-f]{64}$"
        description: Content-addressed digest
    put:
      tags:
      - Blobs
      operationId: putBlob
      summary: Upload blob by digest
      description: Body bytes must hash to the path `digest` (`sha256:<64 hex>`).
        Optional MIME via `Content-Type` or `X-Recension-Mime`. Size limited by `RECENSION_MAX_BLOB_BYTES`
        (default 64 MiB).
      security:
      - ApiKey: []
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
          "*/*":
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Stored
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/BlobPutResponse"
        '400':
          "$ref": "#/components/responses/Error"
        '401':
          "$ref": "#/components/responses/Error"
        '413':
          "$ref": "#/components/responses/Error"
        '500':
          "$ref": "#/components/responses/Error"
        '503':
          "$ref": "#/components/responses/Error"
      parameters:
      - name: X-Recension-Mime
        in: header
        required: false
        schema:
          type: string
        description: Used when Content-Type is missing or application/octet-stream
    head:
      tags:
      - Blobs
      operationId: headBlob
      summary: Check blob exists
      security:
      - ApiKey: []
      - SessionBearer: []
      responses:
        '204':
          description: Blob exists
        '400':
          "$ref": "#/components/responses/Error"
        '401':
          "$ref": "#/components/responses/Error"
        '404':
          description: Not found
        '503':
          "$ref": "#/components/responses/Error"
    get:
      tags:
      - Blobs
      operationId: getBlob
      summary: Download blob
      security:
      - ApiKey: []
      - SessionBearer: []
      responses:
        '200':
          description: Blob body
          headers:
            Content-Type:
              schema:
                type: string
            Content-Length:
              schema:
                type: integer
            X-Recension-Digest:
              schema:
                type: string
              description: Echo of the requested digest
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
            "*/*":
              schema:
                type: string
                format: binary
        '401':
          "$ref": "#/components/responses/Error"
        '404':
          "$ref": "#/components/responses/Error"
        '503':
          "$ref": "#/components/responses/Error"
  "/v1/batch/{team}/{suite}/{batch}/seal":
    parameters:
    - name: team
      in: path
      required: true
      schema:
        "$ref": "#/components/schemas/Slug"
      description: Team slug
    - name: suite
      in: path
      required: true
      schema:
        "$ref": "#/components/schemas/Slug"
      description: Suite slug
    - name: batch
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        description: Batch slug = SDK version/revision
    post:
      tags:
      - Client
      operationId: sealBatch
      summary: Seal a batch
      description: Prevents further submits to this version/revision.
      security:
      - ApiKey: []
      responses:
        '204':
          description: Sealed
        '401':
          "$ref": "#/components/responses/Error"
        '404':
          "$ref": "#/components/responses/Error"
  "/v1/batch/{team}/{suite}/{batch}/promote":
    parameters:
    - name: team
      in: path
      required: true
      schema:
        "$ref": "#/components/schemas/Slug"
      description: Team slug
    - name: suite
      in: path
      required: true
      schema:
        "$ref": "#/components/schemas/Slug"
      description: Suite slug
    - name: batch
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        description: Batch slug = SDK version/revision
    post:
      tags:
      - Batches
      operationId: promoteBatch
      summary: Promote batch as suite baseline (admin+)
      security:
      - SessionBearer: []
      responses:
        '204':
          description: Promoted
        '401':
          "$ref": "#/components/responses/Error"
        '404':
          "$ref": "#/components/responses/Error"
  "/v1/teams":
    get:
      tags:
      - Teams
      operationId: listTeams
      summary: List teams the current user belongs to
      security:
      - SessionBearer: []
      responses:
        '200':
          description: Teams
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Team"
        '401':
          "$ref": "#/components/responses/Error"
    post:
      tags:
      - Teams
      operationId: createTeam
      summary: Create team (caller becomes owner)
      security:
      - SessionBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/EntityRequest"
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Team"
        '400':
          "$ref": "#/components/responses/Error"
        '401':
          "$ref": "#/components/responses/Error"
        '409':
          "$ref": "#/components/responses/Error"
  "/v1/teams/{team}/members":
    parameters:
    - name: team
      in: path
      required: true
      schema:
        "$ref": "#/components/schemas/Slug"
      description: Team slug
    get:
      tags:
      - Teams
      operationId: listMembers
      summary: List team members (viewer+)
      security:
      - SessionBearer: []
      responses:
        '200':
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/TeamMember"
        '401':
          "$ref": "#/components/responses/Error"
        '404':
          "$ref": "#/components/responses/Error"
    post:
      tags:
      - Teams
      operationId: addMember
      summary: Add or update a team member (admin+)
      security:
      - SessionBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/MemberRequest"
      responses:
        '201':
          description: Member added
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TeamMember"
        '400':
          "$ref": "#/components/responses/Error"
        '401':
          "$ref": "#/components/responses/Error"
        '403':
          "$ref": "#/components/responses/Error"
        '404':
          "$ref": "#/components/responses/Error"
  "/v1/teams/{team}/members/{userId}":
    parameters:
    - name: team
      in: path
      required: true
      schema:
        "$ref": "#/components/schemas/Slug"
    - name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    patch:
      tags:
      - Teams
      operationId: updateMember
      summary: Change a member's role (admin+)
      security:
      - SessionBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  "$ref": "#/components/schemas/TeamRole"
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TeamMember"
        '400':
          "$ref": "#/components/responses/Error"
        '401':
          "$ref": "#/components/responses/Error"
        '403':
          "$ref": "#/components/responses/Error"
        '404':
          "$ref": "#/components/responses/Error"
        '409':
          "$ref": "#/components/responses/Error"
    delete:
      tags:
      - Teams
      operationId: removeMember
      summary: Remove a team member (admin+)
      security:
      - SessionBearer: []
      responses:
        '204':
          description: Removed
        '401':
          "$ref": "#/components/responses/Error"
        '403':
          "$ref": "#/components/responses/Error"
        '404':
          "$ref": "#/components/responses/Error"
        '409':
          "$ref": "#/components/responses/Error"
  "/v1/teams/{team}/suites":
    parameters:
    - name: team
      in: path
      required: true
      schema:
        "$ref": "#/components/schemas/Slug"
      description: Team slug
    get:
      tags:
      - Teams
      operationId: listSuites
      summary: List suites for a team (viewer+)
      security:
      - SessionBearer: []
      responses:
        '200':
          description: Suites
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Suite"
        '401':
          "$ref": "#/components/responses/Error"
        '404':
          "$ref": "#/components/responses/Error"
    post:
      tags:
      - Teams
      operationId: createSuite
      summary: Create suite (admin+)
      security:
      - SessionBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/EntityRequest"
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Suite"
        '400':
          "$ref": "#/components/responses/Error"
        '401':
          "$ref": "#/components/responses/Error"
        '404':
          "$ref": "#/components/responses/Error"
        '409':
          "$ref": "#/components/responses/Error"
  "/v1/teams/{team}/suites/{suite}/batches":
    parameters:
    - name: team
      in: path
      required: true
      schema:
        "$ref": "#/components/schemas/Slug"
      description: Team slug
    - name: suite
      in: path
      required: true
      schema:
        "$ref": "#/components/schemas/Slug"
      description: Suite slug
    get:
      tags:
      - Batches
      operationId: listBatches
      summary: List batches with summary counts
      security:
      - SessionBearer: []
      responses:
        '200':
          description: Batch summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/BatchSummary"
        '401':
          "$ref": "#/components/responses/Error"
  "/v1/teams/{team}/suites/{suite}/batches/{batch}":
    parameters:
    - name: team
      in: path
      required: true
      schema:
        "$ref": "#/components/schemas/Slug"
      description: Team slug
    - name: suite
      in: path
      required: true
      schema:
        "$ref": "#/components/schemas/Slug"
      description: Suite slug
    - name: batch
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        description: Batch slug = SDK version/revision
    get:
      tags:
      - Batches
      operationId: getBatch
      summary: Get batch detail with elements
      security:
      - SessionBearer: []
      responses:
        '200':
          description: Batch + element list
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/BatchDetail"
        '401':
          "$ref": "#/components/responses/Error"
        '404':
          "$ref": "#/components/responses/Error"
  "/v1/teams/{team}/suites/{suite}/batches/{batch}/elements/{element}":
    parameters:
    - name: team
      in: path
      required: true
      schema:
        "$ref": "#/components/schemas/Slug"
      description: Team slug
    - name: suite
      in: path
      required: true
      schema:
        "$ref": "#/components/schemas/Slug"
      description: Suite slug
    - name: batch
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        description: Batch slug = SDK version/revision
    - name: element
      in: path
      required: true
      schema:
        type: string
        minLength: 1
      description: Testcase name
    get:
      tags:
      - Batches
      operationId: getElement
      summary: Get one testcase comparison detail (checks, asserts, metrics vs baseline)
      description: |-
        Returns the submitted message plus a comparison against the suite baseline
        (or against another batch when `vs` is set).
        `comparison.metrics` and overview `metricsCount*` / `metricsDuration*` power the
        dashboard duration bars. Requires **viewer+** on the team.
      parameters:
      - name: vs
        in: query
        required: false
        schema:
          type: string
        description: Optional batch slug to compare against instead of the suite baseline
      security:
      - SessionBearer: []
      responses:
        '200':
          description: Element message + comparison
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ElementDetail"
        '401':
          "$ref": "#/components/responses/Error"
        '404':
          "$ref": "#/components/responses/Error"
  "/v1/teams/{team}/invites":
    parameters:
    - name: team
      in: path
      required: true
      schema:
        "$ref": "#/components/schemas/Slug"
    post:
      tags:
      - Teams
      operationId: createInvite
      summary: Create a shareable invite link (admin+)
      security:
      - SessionBearer: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  "$ref": "#/components/schemas/TeamRole"
      responses:
        '201':
          description: Invite created
          content:
            application/json:
              schema:
                type: object
                required: [token, role, expiresAt, path]
                properties:
                  token:
                    type: string
                  role:
                    "$ref": "#/components/schemas/TeamRole"
                  expiresAt:
                    type: string
                    format: date-time
                  path:
                    type: string
                    example: /invite/abc…
        '401':
          "$ref": "#/components/responses/Error"
        '403':
          "$ref": "#/components/responses/Error"
        '404':
          "$ref": "#/components/responses/Error"
  "/v1/invites/{token}":
    parameters:
    - name: token
      in: path
      required: true
      schema:
        type: string
    get:
      tags:
      - Teams
      operationId: getInvite
      summary: Peek invite metadata (public)
      security: []
      responses:
        '200':
          description: Invite info
        '404':
          "$ref": "#/components/responses/Error"
...
components:
  securitySchemes:
    SessionBearer:
      type: http
      scheme: bearer
      bearerFormat: HMAC session token
      description: Dashboard session from login/signup. Default expiry 168h (`RECENSION_SESSION_TTL`).
    ApiKey:
      type: apiKey
      in: header
      name: X-Recension-API-Key
      description: SDK / client key via `X-Recension-API-Key`.
...
  schemas:
    Slug:
      type: string
      minLength: 2
      maxLength: 64
      pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
      description: Lowercase letters, numbers, hyphens
...